updating

+++ src/main/java/ru/timur/clinic/mapper/DoctorMapper.java
@@ -10,10 +10,10 @@ import java.util.List;
 public interface DoctorMapper {
 
     void insertDoctor(Doctor doctor);
-    void deleteDoctorById(Short id);
+    void deleteDoctorById(long id);
     void updateDoctor(Doctor doctor);
     List<Doctor> getDoctorByName(String firstname);
-    List<Doctor> getDoctorById(Short id);
+    List<Doctor> getDoctorById(long id);
     List<Doctor> getAllDoctors();
 
 }
